Added JWT authorization

+++ graphql/resolvers/booking.js
@@ -3,7 +3,11 @@ const User = require("../../models/user");
 const Event = require("../../models/event");
 
 module.exports = {
-  bookings: () => {
+  bookings: (args, req) => {
+    if (!req.isAuthorized) {
+      throw new Error("Unauthorized!");
+    }
+
     return Booking.find()
       .populate(["event", "user"])
       .then((bookings) => {
@@ -16,21 +20,23 @@ module.exports = {
       });
   },
 
-  bookEvent: (args) => {
+  bookEvent: (args, req) => {
+    if (!req.isAuthorized) {
+      throw new Error("Unauthorized!");
+    }
+
     const booking = new Booking({
-      userId: "61f36e47bcfc7dd982b4545d",
+      userId: req.userId,
       event: args.eventId,
     });
     let createdBooking;
     return booking
       .save()
       .then((booking) => {
-        console.log("Entered booking.save()");
         createdBooking = { ...booking._doc };
-        return Promise.all([Event.findById(args.eventId), User.findById("61f36e47bcfc7dd982b4545d")]);
+        return Promise.all([Event.findById(args.eventId), User.findById(req.userId)]);
       })
       .then((res) => {
-        console.log("Entered Promise.all()");
         const event = res[0];
         const user = res[1];
         if (!event) {
@@ -46,7 +52,10 @@ module.exports = {
         throw err;
       });
   },
-  cancelBooking: async (args) => {
+  cancelBooking: async (args, req) => {
+    if (!req.isAuthorized) {
+      throw new Error("Unauthorized!");
+    }
     try {
       const deletedBooking = await Booking.findById(args.bookingId).populate("event");
       await Booking.findByIdAndDelete(args.bookingId);
